smaller modifier bitmask

+++ src/keyboard.hpp
@@ -6,7 +6,7 @@
 
 namespace atmt {
 
-using Modifier = uint32_t;
+using Modifier = uint8_t;
 constexpr Modifier MOD_SHIFT = 1, MOD_ALT = 2, MOD_CTRL = 4, MOD_SUPER = 8,
                    MOD_RIGHT_ALT = 16;
